web/requirements.txt
flask
python-dotenv
flask-cors
gunicorn
kafka-python